pyproject.toml
[project]
name = "insightgape"
version = "0.1.0"
description = "InsightGape: Black Box Auditor - Financial Intelligence Agent (CrewAI multi-agent system for corporate narrative auditing)"
authors = [{ name = "InsightGape Team", email = "insights@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.14.1",
    "rich",
    "python-dotenv",
    "alpha_vantage",
    "pytest",
    "weasyprint",
    "requests",
    "beautifulsoup4"
]

[project.scripts]
insightgape = "insightgape.main:run"
run_crew = "insightgape.main:run"
train = "insightgape.main:train"
replay = "insightgape.main:replay"
test = "insightgape.main:test"
run_with_trigger = "insightgape.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
